Refactor: Remove Search Components and Related Logic
- Deleted SearchInput, SearchResult, Searchbar components and their associated logic.
- Removed search_action.ts and related API actions for website analysis.
- Cleaned up analyze actions and utilities, including data_sync and policy-analyzer.
- Updated the main page layout to reflect the removal of the search components.
- Enhanced utility functions for date formatting and slug generation.
- Adjusted the getSiteAnalysis action to include additional logging and data handling.
